Refactor ChallengeCard component structure and update import paths for consistency
