+++ src/components/groups/challenge/ChallengeCardImage.tsx
