+++ src/components/groups/challenge/ChallengeDescription.tsx
