+++ src/components/groups/challenge/ChallengeHeader.tsx
